Canonicalize D/M pairs generated by Subversion dump streams.
Fixes a bug in the canonicalization code - None placeholders generated
by composition passes were not getting squashed out soon enough, causing
a crash in __ancestor_count.

This revision fails the nut.svn regression - failure left in because it
reveals something that looks like a bug in D/D canonicalization. Look for
data/Makefile.in in the diff output.