Change the algorithm for finding branch comnnections.
Peeviously, we assumed that if you see a directory copy with a given from_rev,
its branch should be rooted at the commit corresponding to that from_rev.
This turns out not to be enough; the symptom is branches can sometimes be
rootedat the wrong place.

What we need to do is drop back to the commit corresponding to the specifed
from_rev, then go back further until we get to a commit on the right branch -
that is, with a commit common path corresponding to the copy from_path.