expunges must work even if they don't delete any events
The expunge command is careful to only delete commits and blobs that
are entirely expunged, leaving behind any commit that included fileops
outside the expunge set, and blobs that are shared with commits not
deleted. However, if there's nothing to delete then it returns early,
failing to keep the newly created repository containing the expunged
operations.