repos without commits shouldn't break anything
It's an edge case, but the expunge command can create a repository
with no commits or blobs, and this thows the rest of the code for a
loop. The Repository object is perfectly valid in this case, but
because it has container emulation methods __len__ and __getitem__
that forward to its events array, it tests as False just like an empty
array would.