Elinimation of all but one find_node() call.
All regression tests pass, except that some M ops are no longer generated
late in the history of the references test.  I am not yet certain this
change is correct, but I don't know that it's wrong either - that repo
has some dubious merges in it.

It seems best to go forward, trusting that code simplification is taking
us in the right direction. Regression tests rebuilt.