use the baton consistently during svn_process
Several of the steps weren't indicated, some of the quickest ones had
counters while longer ones only had the twirling baton or just
nothing. This gives every step a name and every step with a definite
count a counter.

It'd be nice to abstract some of this out, perhaps aiming towards
something like with Counter(list), but that can wait for another time.