Change the way fileops expanded from directory copies are made.
Instead of concatenating them to the end of the array of parsed nodes,
build a new array into which each parsed file node is copied, and
directory copies are expanded into fileops in the new array *as they
occur*.  This won't change the *set* of file ops generated, but it may
change the order in which they will be interpreted at fileop-generation
time.

I did this because the last couple of bugs have been related to mingled
directory copies and fileops on the same directory.  By not reordering
any operations from the sequence in which Subversion's virtual filesystem
would have performed them, I hope to avoid trouble.

One test output changed - references.chk, naturally.  The changes replaced
older blob marks with newer ones, suggesting that this really did fix a
bug that would otherwise have produced stale versions in the translated
history.