Use __slots__ for classes instanciated many times
Each instance of normal classes internally uses a dict to store its
attributes. Use the __slots__ class member of new-style classes to
prevent replace this dict by reserved memory in the instance for every
possible attribute. This saves both memory and access time.

Using __slots__ is beneficial for a class when the class is instanciated
a lot and its instances (mostly) have the same set of attributes. This
is clearly the case for the following classes, sorted by instance count
during an import of a 4GB svn dump of 15000 revisions:
PathMap (917078 instances), NodeAction (417599), FileOp (410454), Blob
(108922), Attribution (15000), RevisionRecord (15000), Commit (15000).

Next are Reset (68) and Tag (15 instances); for now they are kept
unchanged since a repository presumably has much less branches and tags
than commits and files.