Refactor PathMap._find()
Although _find was tail-recursive, it is even shorter and simpler to
read with a reverse loop over components. Also, it avoids repeatedly
poping elements from components which is a slight performance gain.

It will also enable to factor the call to _split_path into _find and
simplify all its callers.

Signed-off-by: Julien "_FrnchFrgg_" RIVAUD <frnchfrgg@free.fr>