Memoize Repository.branchset()
branchset() is mostly used for simple "branch (not) in branchset()" tests
where uniqueness doesn't matter and there is no need to iterate over the
whole set of commits. On the other hand, there are a lot of such tests,
along with some genuine uses of the uniqueness, so there is a need of
optimization across calls. Make branchset() remember its result, like
Repository.commit() does.