Replace sorting throw-away branchlist by sorted()
In two places the code uses .sort() on a newly-created branchlist, only
to iterate over the result. Use the sorted() built-in instead. There is
no efficiency gain (even a slight loss) because Repository.branchlist()
creates a list anyway, but the change will enable Repository to return a
set since these are the only places where a real list was expected.
See http://python3porting.com/improving.html#use-sorted-instead-of-sort