Replace Repository.branchlist() by Repository.branchset()
Now that no caller of branchlist() needs a list, make Repository return
a set instead of converting it to a list. Not counting the set creation
time, this essentially makes all "branch in branchlist()" tests O(1)
instead of O(#branches), and also speeds up a bit cases where the code
iterates over (a sorted) branchlist().