Use an even better parents iteration for mergeinfos
The first rewrite of dump_revprops was simpler, but this one is more
easy to follow since it replaces the iterator slicing by a simple
consumption of its first value.

Also, (parents or []) is only needed to pacify len() and does not need
to be assigned back to parents: when the test is true parents
necessarily remains unchanged.

Lastly, replace the generator expression by a map of str to on the
iterator. This has a smaller interpreter overhead.

Passes the test suite.

Signed-off-by: Julien "_FrnchFrgg_" RIVAUD <frnchfrgg@free.fr>