Replace most uses of dict.items() by dict.iteritems()
When used as the iterator of a for loop, items() is neither Pythonic nor
efficient: it temporarily creates a new list only to iterate over it.
Same for the use of keys(). Replace all occurences of items() but one
with iteritems() and all occurences of keys() but one by iterkeys() or
even nothing since iterating a dict is equivalent to iterkeys().

The two remaining occurences are valid use-cases for items() or keys():
the underlying dict is mutated during the loop and items may be added or
removed. In that case using iteritems() or iterkeys() may raise a
RuntimeError or fail to iterate through all entries. Leave the calls
alone and document why.

See http://docs.python.org/2.7/library/stdtypes.html#dict.iteritems

Signed-off-by: Julien "_FrnchFrgg_" RIVAUD <frnchfrgg@free.fr>