Replace result = [iterator][0] by result = next(iterator)
[x for x in some_list if predicate][0] builds the entire list, filtering
some_list from end to end, to instantly discard all matches but the
first. Use next() on a generator expression instead, which only searches
until the first match and breaks early.

Note that this will raise the StopIteration exception if not found,
instead of IndexError for the previous code. This is not likely to
matter though, since the exception wasn't caught either.

Signed-off-by: Julien "_FrnchFrgg_" RIVAUD <frnchfrgg@free.fr>