Do not share toplevel pathmaps
Change the PathMap logic so that toplevel PathMap instances never share
their dictionnary. This is needed for later refactoring where the
decision and action to unshare will be undertaken by the parent PathMap.

Note that the change is unlikely to have a big memory impact since most
snapshots asked by callers external to PathMap are instantly written to,
thus unshared.