Share PathMap instances instead of just their dict
Now that toplevel PathMaps never have shared status, everytime a PathMap
with shared |elements| is instanciated, it is from a parent PathMap.
Make this parent PathMap never create the instance but keep the old
PathMap while using its |shared| attribute to mean that the PathMap
itself is shared, not only its dict.
This enables opportunistic unsharing, by replacing as needed at _insert
or _delete time shared PathMap children by newly created "unshared"
shapshots. Again this is possible because at that time we have access to
both the parent and the shared child.

Note that _unshare is now unconditionnal.