Add a more efficient storage for slowly changing mappings
Introduce the |evolvingdict| class, which is a dict-like structure
optimized for the PathMap use-case of a series of dicts whose values
are mostly unchanged from one to the next. Implement only the interface
needed for PathMap to use it.

Each key maps to a list of successive values indexed by an unique
integer |snapid| which gets incremented for each new snapshot. These
lists are not enlarged at the time of the snapshot, but as needed at the
next assignment.

In the future the lists can be made "sparse" by only storing (<snapid of
the change>, <new value>) pairs. Keep things simple for now.