Don't choke on out-of-range mergeinfo
If a mergeinfo property contains impossible information, like a 6-digit
rev when the repository only goes to 5-digit revisions, the current code
aborts with a KeyError. Skip the mergeinfo instead. There may be more
useful resolutions, like clamping to plausible range, but a predictable
tool is better than one trying to be smart and sometimes failing with
really stupid output.