Dont mistake branches with "tags" in their name for real tags
The commit creating an SVN branch/tag is changed into a Git-like tag
whose name is the SVN branch/tag name, as is for tags and with "-root"
appended for branches. The test was implemented by searching "tags" into
the branch name, which means "branches/extagsy" would not get the
"-root" treatment. Change the check to a test whether "tags/" is a
prefix.