Revert most of cd04bd02c224a4e03fadbc393646cfe0359b568b
That commit goal was to make a future transition to Python3 the least
painful possible by using the 2to3 utility of Python, but its action is
actually detrimental if you don't switch the interpreter to Python3.
Moreover 2to3 isn't able to cope with being run several times on the
same code, and here none of the changes were meaningful. Revert most of
the commit (the popen_or_die case is be handled in another commit).