Revert popen_or_die raise to simple Python2
cd04bd02c made this code utterly meaningless, but in fact it was
uneffective since it was added by 51a79c062 : the Python2 syntax raises
a SyntaxError in Python3 *at parse time*, and thus cannot be caught by
the code. The only usable way would to wrap it in an exec() call so that
the parsing of the offending line is delayed until it can be caught.
Since the rest of the code isn't Python3 ready anyway, simplify our
lives by using the old syntax unconditionnally.