Revert f99701c273
2to3 should be used at the time of transition to Python3, not when
continuing to use Python2 because its changes can be detrimental
performance-wise, and even alter the behaviour of the program. An
example is the __nonzero__ -> __bool__ change where __bool__ is not
recognized as special by Python2 and suddenly the stock and dumb Python
True/False conversion kicks in (since the tests passed it seems that
this function is unused by the code, but...).

Revert the whole commit except for a case which will be handled
differently.