Use pairwise iteration instead of a temporary list
Instead of creating a list so that two successive commits can be
random-accessed in an iteration over indices, use a recipe from
itertools to simultaneously iterate through two iterators while
the second is one item ahead from the first. This avoids a big list
creation when there are a lot of revisions.