Prefer as_string() to str() for ReposurgeonEmail's str()
email.message.Message.__str__ is equivalent to as_string(unixfrom=True)
in Python 2, but apparently to as_string(unixfrom=False) in Python 3.
This makes mailbox roundtripping fail. Use the explicit form for more
control, and to be Python3-proof. While we're at it, use super() instead
of hard-coding the next class in the Method Resolution Order; this needs
RepoSurgeonEmail to be a new-style class.