Replace cmp sort arg by modern equivalent
The cmp arg of sort() or sorted() is deprecated and raises an error in
Python 3 onward. Since FileOps only are ever sorted with |compare| as a
cmp method, replace this mechanism by making the FileOps themselves have
a total ordering. Use the __lt__ rich comparison method for that
purpose; this also makes the code clearer.