Better fix for files touched more than once in the same SVN rev
Before 29ac14817c95c03bfc4a33cdd83ea8a4027dca10 "Fix and regression test
for yet another Subversion edge case", when a SVN revision modified a
file in more than one expanded node, the code did not consider the
former as a possible ancestor of the latter and searched instead in
previons revisions.

That commit changed the logic to first loop over the previous expanded
nodes of the current revision, but the loop chooses the first one
touching that path instead of the last, fails to consider "change"
actions as valid ancestors, and doesn't cater for the case where a
"delete" follows the "add" (then the "add" is not a valid ancestor).
Moreover, using an inner loop makes expanded nodes conversion O(n^2)
where n is the number of expanded nodes in the revision.

Try another method by remembering the current revision's expanded nodes
so far, indexed by their touched path, and trying them before falling
back on the filemaps of previous revisions. This has several advantages:

- It doesn't duplicate the ancestor finding logic but rather relies on
  it. This means a lighter maintainance burden when the ancestor logic
  needs to be tweaked.

- It is a small and simple change.

- It only multiplies the expanded nodes parsing time by the amortized
  cost of an insertion and a search in an hash table, which is O(1).

The drawback is that it uses a bit more memory: it stores n nodes with
their paths in an hash table. That memory is reclaimed as soon as the
next revision is handled so it shouldn't be a problem.

Remark: That bug was responsible for an assertion on node.blobmark being
null in the following way (which exists in the wild) :

r1: a dir "trunk/x" is created with a file "a" in it.
r2: the dir is copied to "trunk/y", then the file "trunk/y/a" has a
    "change" action on it. This node has no blob (a "change" without
    Text-content-length is strange but seems valid in subversion land),
    so reposurgeon uses the blobmark of an ancestor which it couldn't
    find in r1. No assertion here, the node is skipped since the node
    has neither an ancestor nor a blob.
r3: "trunk/y/a" is copied to "trunk/z/a", so gets its blobmark, and the
    assertion is triggered.

This is essentially what is done by the test added in commit
29ac14817c95c03bfc4a33cdd83ea8a4027dca10.