Replace usage of set() by a simpler direct test
The branchlinks operation was creating a set of commits satisfying a
predicate to later test if some commit.branch are in the set. This is
overkill since the predicate only depends on the one commit and not on
other complex intertwinned conditions.

Replace the mechanism by a simple predicate test on commit.branch. The
test is made able to cater for the case where commit.branch is not a
commit and has no fileops member.

Also remove the use of a temporary variable used only once.