add_parent: don't forget newparent._child_nodes, update it
If newparent already has its children cache, it is O(1) to update it to
reflect the new situation. Do so instead of clobbering it and forcing a
costly recompute on next children() call.

Note there is a semantic change: if add_parent is called multiple times
with the same parent, then it will appear multiple times in the updated
children cache. A check could have been added to avoid that but:
 - it makes the update O(#children) instead of O(1)
 - it breaks the duality between children and parents since the
   duplicate is kept in the latter.

Due to the way cache regeneration works, the duplicates will be lost on
cache deletion.