remove_parent: update newparent._child_nodes
If newparent already has its children cache, it is O(#children) to
update it to reflect the new situation. Do so instead of clobbering it
and forcing a costly recompute on next children() call.

Due to the way cache regeneration works, in the old code all occurences
of self disappear from event.chidldren(). Since callers rely on that
fact, ensure that it still is the case by filtering the whole list
instead of using remove().