replace_parent: update children cache
Updating the children cache of e1 is O(#children), and it is O(1) for
e2. Prefer that over erasing.

Here only one occurence of self is removed from e1's cache since it is
more fitted to the semantics of replace_parent(), and no caller expect
more than that anyway.