clone: don't short-circuit the parent encapsulation
By doing a simple structural copy, Commit.clone() created a commit whose
parents are not aware of. It didn't matter when the children list
was generated on demand, but now that the parent-related method try to
stay on top of the children bookkeeping one can no longer afford to
meddle with the encapsulation internal state.

Use set_parents() to reset the parent list through normal means.