Make pylint 0.26 happy by disabling a test and fixing another
pylint 0.26 added some new tests, two of which trigger errors that are
currently best left ignored. These are:

[R0924, PathMap] Badly implemented Container, implements __delitem__,
__getitem__, __setitem__ but not __len__

[R0924, Repository] Badly implemented Container, implements __getitem__,
__len__, __setitem__ but not __delitem__

and

[W1401] Anomalous backslash in string: '\0'. String constant might be
missing an r prefix.

Silence the first two by adding the missing methods. Disable the last
for now: it will be tackled by the string audit prior to porting to
Python 3.