svn_process: refactor cvs2svn D/M pairs removal
For each DELETE expanded node, the code created the slice of later nodes
then iterated over it to find one with the same path. This method was
instanciating a lot of temporary lists, and was globally O(d*n) where d
is the number of delete nodes and n the number of nodes. Replace the
inner loop by a set lookup for already seen paths while the outer loop
is made to iterate in the reverse order.

Note that since a lot of expanded nodes can be generated by subdir
copies or deletions, the cost was high. Changing that single loop from
O(d*n) to O(n) cut by four the commits phase duration on my test
repository...