remember: accept a target argument in place of a committish
When callers of remember() already have the underlying commit object,
passing the mark to remember() is wasteful since it then searches the
commit in the event list. Make remember accept either a committish
argument as before, or a target argument from which the mark is then
obtained. If both target and committish are given, the target takes
precedence.