Refactor per-file properties handling.
The code was complicated, the use of latch was making unclear what the
code did on which conditions. In particular:

If global_options["svn_ignore_properties"] was True, latch wasn't ever
set so node.path was deleted without further ado if it had ever been
set.  The whole code would then be a big churning no-op. The main
purpose of latch is in fact to detect the first prop passing the tests,
and remember if there was any.

Make the intent clearer by:
 - conditionning all the block on "svn_ignore_properties";
 - using a standard python recipe to detect whether an iterator has
   elements or not without costly operations.

Also, since self.has_properties was not used anywhere else, make it a
local variable instead.