Fix the ordering of split commits
The comments said that deleteall cliques were moved last by the sort,
but "not b[0] in deletealls" gives 0 for deletealls and 1 for others,
thus actually putting deletealls in front. Rather than fixing the
comment, make the code do what is announced because it seems more
sensible especially in the non-mixed case where the base commit with
unmodified fossil_id should be the one with interesting operations.

Also apply the needed changes to the testsuite expectations.