Rework the oplist generation for split commits
Instead of creating the set of branches only concerned by deletealls,
then sorting based on being in that set or not, just conditionnally
append during the loop to create one deleteall-only list, and one list
for meaningful operations. This avoids a sort, multiple set additions
and lookups, and makes the code more readable. In particular it is now
obvious that deletealls are indeed put last.