Avoid repeatedly enumerating all repo events
Getting the list of all repository commits to iterate it in reverse
costs much more than reverse iterating over all events and manually
filter non commits. This was already the case before the migration
of commits() from list to iterator but it is even more blatant now.

Use the faster method and add a comment explaining why.