Speed up the gathering of actions into cliques
The operation is O(k*b) where k is the number of actions and b is the
number of branches. Add an heuristic trial of the last used branch
before searching for other branches. This undercuts the "b" factor to
make the operation O(k+b) when all actions are on the same branch, which
is the most frequent case by far.

This change represents a -80% evolution of the total time taken by this
phase over my test SVN dump. This also meant -20% on the total commits
phase (66s to 54s) thus -9% on the total reposurgeon run time.