PathMap: Speed up splitting a path into components
Instead of doing multiple consecutive calls to os.path.split, normalize
the path once and use a textual split on the os.sep separator, while
still ignoring empty components. This is simpler and faster, avoids
extracting the components in the reverse order and should be as robust
as the previous version.

This represents a 5% speedup on the filemaps phase and another 5% on the
commits phase. Presumably other phases benefit from it too but the
difference is not noticeable.