Fix parsing of recursive SVN delete with svn:ignore properties at multiple levels
When deleting a directory, the SVN parser currently only checks the deleted
directory itself for active .gitignore files generated from svn:ignore properties,
which can create a problem for recursive deletes. Consider the following tree:

- a
  - .gitignore
  - b
    - .gitignore

When removing a, the parser correctly creates an event that deletes a/.gitignore,
but it won't delete a/b/.gitignore. The a/b hierarchy thus never gets cleaned up.

This patch fixes the problem by scanning all active .gitignore files and creating
delete events for every entry within the directory hierarchy that is being deleted.