Replace FileOp comparison with a suitable key
When sorting, the class comparison special methods get called a lot and
can end up as bottlenecks. A key function is instead garanteed to be
called only once per item to sort. Use the latter to meaningfully
replace the former.

This change reduces a lot that part of the code and increases the
signal-to-noise ratio. This also represents a 2-3% gain on a typical
repository, and could be much more on a repository with commits touching
a lot of files.