Make paths correctly handle redirection
The test for rest being empty was not a good one since the redirection
part is still in it. Also, when passing the full line to LineParse one
can confuse it with angle branckets from tag/branch/fossils sets.

Test for presence of sub or sup, and pass only the end of the line to
LineParse.