mailbox_in: Ensure event is current and fix error reporting
If event as obtained by the previous search had no email_in attribute,
reposurgeon complained, using eventnum to refer to the event even if the
code path taken by the search phase didn't populate it. Try to use
event.index() instead with a fallback on a numberless message.

Also set event to None at the beginning of the search phase to ensure
that it is not a leftover of a previous update when no sense can be made
of the current one.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>