Commit.delete: don't normalize or coalesce deleted commits
The code was adding commits to the altered list to later normalize their
fileop list, and did so for kept and deleted commits alike. This could
lead to problemr since the final fileop list of a deleted commit can
depend on deletion order.

Skip deleted commits in the normalize/coalesce phase to avoid spurious
warnings or any dependency on the order of deletion.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>