eval_pathset: Add a flag to match on any path in tree
pathset matched only for commits with a fileop touching the file. Make
it understand a leading '*' flag telling reposurgeon to search for paths
in the full file tree in the commit.

Though the implementation could have used Commit.manifest(), it prefers
an ad-hoc search which is less memory and CPU hungry.

Also add tests, and rewrite documentation to better express the
interactions between '*', paths and regexes.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>