eval_pathset_full: use marks as keys, not commits
Since the __hash__ function of Commit has not been overloaded, the hash
is by default the id() of the instance, that is its memory adress. If a
Commit object is relocated during the selection loop, its hash will
change and sets or dicts will raise a KeyError when attempting to use it
as a key. Use the commit mark instead since it is guaranteed not to
change during the loop.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>