Repository.delete: Reparent all children for sure
Since the children optimization series, the list returned by
Commit.children() is not a new immutable list anymore but changes as the
children come and go. In particular, the reparenting loop was skipping
some children because the list was mutated underneath it.

Take a copy of the list prior to looping over it and emptying it.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>