Repository.delete: respect first parent when reparenting
To reparent children of a deleted commit to its parents, delete was
first removing the unwanted parent, then appended back the new parents.
That had the side effect of changing the order of parents.

In particular, if the deleted commit was the first parent, the second
parent suddenly was promoted to first parent which is wrong in most
cases. Indeed, the tree of a commit is obtained by inheriting its *first
parent* tree then applying the fileops (see the fast-import format
documentation after which are modeled Reposurgeon internals).

To ensure correctness of the trees, one has to ensure that if the commit
currently deleted, say "a", is the first parent of one of its children,
say "b", then the first parent of "b" is replaced by that of "a".
Because the fileops of "a" will be pushed either to "b" or to its new
parent, its implied tree will then stay the same (unless obliterate is
used of course).

Use a slightly stronger behaviour by always putting the first parent of
the deleted commit in the same place as the latter in the children
parent lists.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>