Use a more pythonic idiom to find first "root" commit
Previous code used a for loop with a break when the wanted commit was
found; use the next() built-in on a generator expression instead. This
is faster since the loop is run in C-speed code and more explicit.

Also, instead of checking first if "root" is in branches, rely on next()
raising StopIteration when no commit is found on the "root" branch.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>