svn_process: use marks as keys for self.branchlink
The hash of a Commit object is its memory address and can thus change
unexpectedly. This makes such objects bad candidates for dict keys. Use
the commit mark instead but store the commit object in the value to
avoid having only its mark and needing to search for it later.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>