Make PathMap able to store data on directories
When setting a value with a non-string path, its first item is extracted
as the path and a special child of the path is written to instead of
writing to the path itself and clobbering the directory structure that
might be present underneath.

This enables attaching any non-None object on directories, by having the
same effect as appending a component to the path that is certain never
to be used by any content put in the PathMap() by any other mean.

These special entries are skipped over when enumerating.