branchset: return all branch names ever created.
Resets with a committish create a branch even if there are no further
commits to it. Take this case into account.

Note that this doesn't correspond to the branches that git will see:
whereas git forgets branches when a reset without committish is
encountered, branchset() does not because callers want branches that
exist at some point, not just at the end of the stream.