branchmap: Take resets into account
The previous code forgot to update the branch mapping with the action of
reset commands in the stream. Apply git fast-import rules instead.

Note that since the method returns *terminal* marks of any branch, when
a reset is encountered that has no committish the branch is completely
forgotten, which corresponds to what git does. In other words, the
returned directory is a map from branch names to commit marks of their
tip *at the end of the stream*. In particular, set(branchmap()) is
included in, but not always equal to branchset().