Get the ancestor with from_path before from_hash
Previous code used from_hash to retrieve the ancestor node of a copy,
but that only returns the first node encountered with the wanted
content_hash. Its path and metadata might be otherwise unrelated.  First
try to fetch with from_path and from_rev instead, then fallback on
from_hash.