svn_process: fix interaction of copies and ignores
While SVN copies the svn:ignore property of dirs while copying them to
their new destination, reposurgeon completely ignored them, be it for
new branch creation, or for other copies for which it generated subnodes.

In the first case there is no need to actually emit fileops for the
.gitignore files because they get inherited from the first parent.
In the second, re-create all .gitignore files in the destination from
the files of the source.

In both cases, remember that we now have new active .gitignore files so
that we correctly emit D fileops when directories are removed.

This makes our generated .gitignore files mirror svn:ignore properties
of SVN nodes in all cases.