svn_process: prepend default ignores more often
reposurgeon only included default SVN ignore list for the trunk branch,
for an SD_ADD node, and if there was no property at all on it. To better
simulate Subversion behaviour, prepend the default list to the toplevel
svn:ignore property of all branches. This list will end up in the
.gitignore file of the tree root in all generated git branches.

This change also reduces mismatches between branches/tags and trunk
ignores and thus reduces churn on .gitignore files.