svn_process: prefer tagify_empty to ad-noc local code
This factorizes most of the logic out of svn_process and makes it more
robust (especially when there are multiple tipdeletes on the same
branch). Note that since tagification of branch roots is later in the
process, the CVS junk cleanup phase has more chances to clean branch
roots, this replaces some Tags by Resets in the test suite.