svn_process: don't create links to deleted commits
When svn_process tagifies the root commit, it doesn't check that it is
not the parent member of a tuple in branchlinks. The branch link
creation pass happily obeys and now some commits can have a parent whose
self.repo is None, which breaks a lot of things elsewhere (the test
suite happens to somehow carefully tiptoe around the breakage).

Fix the problem by ignoring commits not attached to a repository when
creating the links. The branchify test changes.