expunge: ensure parents are in the same repository
The expunge code filtered parents in cloned Commits to only keep Commits
whose mark was in the expunged repository. But this only made the
parents to point to the original commits with the same mark in the
non-expunged repository.

Use set_parent_marks() instead to force a lookup of expunged Commits
from their mark. This changes two tests.