Repository.delete: Don't use Commits as set members
Since the hash of an object without custom __hash__ is based on its
memory location, it is not stable which can cause sets and dicts to
search for the object in the wrong hash table slot and raise KeyError
even if the object is in it. Pypy might even outright crash in that
case. Use the deletehook member instead for the same purpose.