Repository.delete: Fix lightweight tags conservation
delete() was setting the branch of the new target to the deleted event
branch. The code had several issues:

* it didn't use the new_target computed just below, which meant
   duplication of the logic.
 * it checked multiple times if the event had children, even when it was
   already known that it is the case. Similarly it computed identical
   even when the result was already known (because it just set the
   child's branch to event.branch)
 * The choice of clobbering the new target branch meant that a
   lightweight tag on the new target might be destroyed along the way,
   or its branch misplaced (or even lost) if the target was the last (or
   only) commit of its branch.

Instead, just replace the commit by a Reset which will do nothing more
than move the branch to the right point. The attachments reparenting
will then take care of pointing that new Reset to the correct target.

Also modify the test checking if all tags (and target branch) are kept.