svn_process: ignore single-rev mergeinfo
svn:mergeinfo fields that consist of a single rev come from |svn merge
-r xxxx| commands where xxxx is a single revision: this is a cherry-pick
operation and inferring a merge link is wrong in that case. Don't do it.

The expected results split.chk and mergeinfo.chk changed as a result,
and the test mergeinfo-with-split was updated to continue testing merge
generation with split commits.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>