Make delete handle pushback with no parents
delete() crashed with a value error if the pushback policy was in use
and the deleted commit had no parent. Emit a warning instead. While
we're at it, warn the user if he deletes a commit without children and
without obliterate policy nor pushback.

A test changes because it triggers the new warning, and another test
changes because of a break removal; it prevented the new checks to be
done on each selected commits and it is interesting to have all warnings
anyway.