don't use self.selection in Reposurgeon.edit()
edit() is supposed to get the selection set through its selection
argument, yet used self.selection() for the single Blob special case.
The error was side-stepped because the only caller of edit() that cared
about blobs had selection == self.selection, but fix it nonetheless.

Also make clear that a precondition for edit() is that self.chosen() is
not None, and selection musn't be None either.