Make do_unmerge() robust against no selection
The unpacking of self.selection was already in a try/except block, but
that block didn't catch TypeError which is raised when unpacking None
(ValueError is raised when self.selection is iterable but not of the
correct length). Also, make self.selected() responsible for fetching
the selected event if it is a Commit.