Don't renumber in absorb
The absorbed repository is already uniquified. So renumbering will put
the absorbed repo back into the mark-namespace of the absorbing repo.
Even if the absorbing repo would be correctly renumbered (this is
currently no precondition) the renumbering has an off-by-one error.
The new graft2 test shows the behaviour. With the old code we end up
with a parent reference to a blob. Bad.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>