eval_pathset_full: Remove dead code
Due to the v3 parser changes, litteral path matching can no longer take
flags. In particular there is no mean to tell it to search into
manifests, and eval_pathset_full is never called in the non-regex case.

Remove the check for a callable match_condition, and all code that
caters for litteral paths in eval_pathset_full.