Better implementation of @dsc
dsc_handler() was essentially doing a breadth-first traversal of the
descendants by stages: reached commits were not directly added to an
“open” set but rather to a “beyond” set, which replaced the “open” set
when the latter was empty. Prefer a simple “fist-in, first-out” queue so
that the implementation is more in line with classic BFS.

Furthermore, the traversal never checked whether a commit had already
been reached (the “closed” set in BFS terms) and only relied on
duplicates removal by the Python set type used to store the “beyond”
commits. A given commit that could be reached from the selection set by
more that one chain of descendants was thus considered multiple times
along with every descendant, unless all of the said chains were of the
same length --- so that the commit is never reached in two different
stages. Unconditionally check if the new commit is already in the found
descendants, so as not to queue it anew.

This should be a big improvement in memory usage and computation time
for histories with a lot of merges --- with no merge point the ancestry
graph is a tree and the problem vanishes.