Makefile: generalize cyreposurgeon build for platform variations
As Python 3 is default on Arch, they change the executable names of
cython to cython2 for the Python 2 version; Debian takes the different
direction by using cython3 as the name for the Python 3 version.
Change the Makefile so that it has a CYTHON variable that can be
changed at build time.

Another Debian-ism is the plain "python" package name for pkg-config,
on Arch this does not even exist.  However, python-2.7 is universal,
so let's just use that.

Additionally the C compiler variables are removed; these are
automatically set by Make anyway, and can always be changed at build
time.