Improved Hg test. It nearly round-trips now
testrepo2.fi is the result of running testrepo.fi through fi-to-fi;
thereby causing the canonicalisation rules to be applied.  If we
just test with testrepo.fi we get mismatches when reposurgeon
canonicalises, but since canonicalisation is idempotent, using
testrepo2.fi should work.

Also we had to set "sourcetype git" in hg-to-fi to prevent hg-style
 nl-after-comment addition.

Remaining failure is that we get an extra
"reset refs/heads/master"
just before the first
"commit refs/heads/master"