invert logic of eval_disjunct()
Instead of removing selected events from the set of unselected events
and then removing that result from all events to determine the set of
selected events, simply build up a set of selected events and return
that.

In addition to improving readability, this change will make it
possible to use an ordered set to preserve the user-specified order of
selected events.