Fix Gitlab issue #35: mailbox_out add a full stop before dash
This was a hangover from when the test for message delimiter in the
mailbox was much looser.  It was intrended to protect text lines that
were false matches for the delimiter, it seems I never got that code
right. The news version should work better and make false mmatches
extremely unlikely.