fix selection processing in eval_disjunct()
Before, the | operator did not perform a simple set union operation as
documented.  Instead, it did something like this:

  * of all events, find those that match the left-hand expression
  * of the remaining events, find those that match the right-hand
    expression
  * take the union of the two resulting sets

The "of the remaining events" part makes a difference for expressions
like @max(=C) in the right-hand expression.  For example,
@max(=C)|@max(=C)|@max(=C) should behave the same as @max(=C) (select
the last commit), but because of the bug it actually selected the
last, second-to-last, and third-to-last commits.

Now eval_disjunct() behaves like this:

  * of all events, find those that match the left-hand expression
  * of all events, find those that match the right-hand expression
  * take the union of the two resulting sets

which matches the documented behavior.

Note:  The 'unselected' variable is still used to construct this
function's return value, so the 'unselected = set(preselection)' line
is still needed.  This function's logic could be tweaked to eliminate
that 'unselected' variable and improve readability, but I will address
that in a separate commit.