fix selection processing in eval_conjunct()
Before, the & operator did not perform a simple set intersection
operation as documented.  Instead, it did something like this:

  * of all events, find those that match the left-hand expression
  * of those events, find the ones that match the right-hand
    expression
  * take the intersection of the two resulting sets

The "of those events" part makes a difference for relative expressions
like @max(=C) in the right-hand expression.  For example,
@min(=C)&@max(=C) should be the empty set unless there is only one
commit, but because of the bug it actually selected the first commit.

Now eval_conjunct() behaves like this:

  * of all events, find those that match the left-hand expression
  * of all events, find those that match the right-hand expression
  * take the intersection of the two resulting sets

which matches the documented behavior.