Forced mercurial to santize author names.
In mercurial, author names can be almost anything. If the name
didn't have an email, it would fail to load. This fixes that by
forcing mercurial to try to find the parts and load them into
a given format. It's likely that emails it makes up will be wrong,
but it does a good enough job that a well-done authormap can
finish the job, if it gets the chance.

This does break the unit tests, but only because it changes the
expectation of preservation on the quotes. The fast-export format
doesn't care, but it could cause problems somewhere, if those quotes
were expected.