Made reposurgeon more resiliant against subrepos.
Mercurial subrepositories may exist in historical changesets that cannot
be cloned. This prevents mercurial from ever updating to that revision
again. This is a terrible workaround that fakes updates in such a way as
to ignore the busted subrepos. This is always safe because the subrepos
aren't included in the repository conversion and never have been, even
if they're still valid and accessible.