SubversionDumper: avoid incorrectly removing directories from pathmap
Assuming a repo with structure:
trunk/dir1/dir2/file1
trunk/dir1/dir2/file2
and a commit that deletes file1.
Before this change dir1 is removed from the pathmap when processing the deletion even though dir2 is not.  If file2 is then deleted a traceback occurs.  After this change dir1 is not removed from the pathmap and no traceback occurs when file2 is deleted.