SubversionDumper: improve SVN branch detection
Before this commit the code struggled with simple branched repositiories especially
when a commit at or prior to the branch point was tagged.  The logic introduced by
this commit to assign a branch to a commit is rather rudimentary.

Test altered as now a bogus tag is no longer created.