Fix 'filter --shell' command not being invoked for commits
The shell command provided to 'filter --shell' is never invoked for
commits because the 'attributes' set it passes to data_traverse() is
unconditionally empty, so data_traverse() never even attempts to apply
the filter to any fields of the commit.

This was broken by <2014-04-17T07:07:49-04:00!esr@thyrsus.com>, which
also broke 'filter --dedos' due to the same cause, however, --dedos
was fixed by <2014-04-17T07:14:25-04:00!esr@thyrsus.com>, but --shell
never was, so fix it now.

The breakage went unnoticed because bogus test output (test/filter.chk)
was committed along with the bug.

In the long run, a more robust approach might be for data_traverse()
to default to the set {"c", "a", "C"} if the provided set is empty or
None, however, that is outside the scope of this bug fix.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>