Fix crash when 'mailbox_in' adds an author
When Commit.email_in() detects that the incoming mailbox message adds
one or more new authors to a commit, it expands the authors[] array,
initializing each new element with a virgin Attribution in which the
'date' property is None. When it later goes to add the new author, it
attempts to compare the incoming date against the date in the virgin
Attribution and crashes when Date.__ne__ tries accessing None.timestamp.

A similar treatment was applied to the committer date in
<2012-01-03T23:56:12-05:00!esr@thyrsus.com>.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>