repomapper: revert accidental Python 3-only restriction
repomapper is intended to work with both Python 2 and Python 3. Its own
source code contains the comment:

    This code runs under both Python 2 and Python 3: preserve this
    property!

and it does indeed work fine with Python 2. However,
<2016-02-28T07:49:08-05:00!esr@thyrsus.com> made a change (likely
detritus from a debugging session) forcing it to use only Python 3,
which is problematic for platforms (such as stock MacOS) lacking
Python 3. Revert this.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>