test/setpython: detect Python installation location more robustly
The 'setpython' scripts needs to know the installation location of the
requested Python, and it discovers it with 'whereis'. However, the
'whereis' command on MacOS checks only a restricted set of directories
for the queried command. It does not, for instance, check
/usr/local/bin, which is a typical installation location for programs
built manually or installed via Homebrew. MacPorts installs programs in
/opt/local/bin. Neither of these locations is searched by MacOS
'whereis'.

Work around this shortcoming by employing a more robust method of
determining executable location.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>